Add queue name and target to admin UI
Summary:
Since I'm going to be changing the way that targets are propagated in the
future, this makes it easier to verify that things are working correctly (and
notice problems with the existing system).

Test Plan:
Load a failed pipeline and a successful pipeline and click around to look at the
different child pipelines. All of them display the correct queue name and
target.

Reviewers: mattfaus, chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D8556