Handle the situation when a Barrier that relies on a Slot is missing.
If there are task and pipeline kickoff retries it's possible for a _BarrierIndex
to exist for a _BarrierRecord that was not successfully written. It's safe to
ignore this because the original task that wrote the _BarrierIndex and
_BarrierRecord would not have made progress to kick off a real pipeline or child
pipeline unless all of the writes for these dependent entities went through. We
assume that the instigator retried from scratch and somehwere there exists a
good _BarrierIndex and corresponding _BarrierRecord that tries to accomplish the
same thing.

R=ttrimble
DELTA=35 (34 added, 0 deleted, 1 changed)

Revision created by MOE tool push_codebase.
MOE_MIGRATION=7281