Fix the "next page" cursor link on the root pipelines page.
Before this change it would drop the "class_path" parameter so you'd only see the next page for *all* pipelines, not the one you already selected. This required some tricky string manipulation because the two RPCs to the server side (one to list roots, another to list pipelines) come out of order. An alternative approach is to parse the query string, but I'm trying to avoid writing that code again or pulling in a dependency on a library that does it for me.

Also fixes the content-type on the JSON return value. It needs to be application/json to make sure any middleware doesn't try to do manipulations to the data as if it were JavaScript (like inserting comments!).

R=ttrimble
DELTA=11 (8 added, 0 deleted, 3 changed)

Revision created by MOE tool push_codebase.
MOE_MIGRATION=7282