Missing line from Pull Request #19
A bad copy/paste job.

Also wanted to mention that this has nothing to do with Issue #17 - a misread on my part. It appears as if the pipeline library will already use the same queue name in which the executing task is part of - my interpretation says this should be the parent pipeline so maybe an example on how issue #17.

What PR #19 (and this) does is fixes the situation in which child pipelines were given a target to execute to but that target was not propagated when queueing the tasks.