Fixes Issue #60
This was tried differently before in commit 443e311.

For barrier notifications, this was tried before by getting all barrier records' pipeline keys and doing a batch get, but that method did not account for situations in which barrier checks were fired but no pipes were ready yet. My method will potentially result in more RPC requests, but only pipes that are ready to be executed will be fetched. 

For retries: we already have the `pipeline_record` fetched, just pass in the target parameter